Changed meeting_id to be optional (FE + BE)
The UUID was being generated in the BE and DB which was more checks than necessary. It felt more appropriate to let the DB take care of that. The FE was modified the inform the user that they did not have to fill in the ID field if they did not want to.