Split the db seed file into 5 seed files
Beyond the obvious 4, I added the linkage seed file. This separate DB will be another microservice that manages the many-to-many connections between the different tables

Obviously, this part needs a lot more work.

I also had to remove all foreign keys from the DB configuration. Lots of changes will have to be made the backend layer to account for this change.