Updated service.py and main.py
Added a service that takes the data as input, and sends HTTP requests (for the time being; will be adjusted once EDA is up and running) for each of the meetings in the data packet, call the service function (new one) for sending all the requests related to the batches. and an error counter to make sure the results are hitting the correct error rate.